requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
opencv-python>=4.9
numpy>=1.26
anthropic>=0.40
openai>=1.40
requests>=2.31
